requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
openai==1.12.0
jinja2==3.1.3
pydantic-settings==2.1.0
python-multipart==0.0.9
